app/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.39.0
httpx==0.27.2
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
jsonschema==4.20.0
